Add a zsh framework-compatible plugin file
Add a plugin file that is compatible with the ZSH plugin standard (see https://wiki.zshell.dev/community/zsh_plugin_standard).

This allows ZSH frameworks to automatically load pay-respects if it is
instlled on a machine.

The plugin file will only be used by ZSH frameworks and will not
interfere when people are not using one.

Signed-off-by: Joe Block <jpb@unixorn.net>